# swagger.yaml
openapi: 3.0.0
info:
  title: 'PDB API'
  description: 'API documentation.'
  version: '1.0'
servers:
  -
    url: 'https://pdb.local'
    description: Dev
  -
    url: 'https://localhost'
    description: Local
paths:
...
  /command/factory:
    post:
      tags:
        - command
        - factory
      description: 'Create a new factory.'
      operationId: createFactory
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FactoryCreateRequest'
      responses:
        '201':
          $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/schemas/ErrorResponse'
  '/command/factory/{id}':
    put:
      tags:
        - command
        - factory
      description: 'Update a factory.'
      operationId: updateFactory
      parameters:
        -
          name: id
          in: path
          description: 'Factory ID'
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FactoryUpdateRequest'
      responses:
        '200':
          $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/schemas/ErrorResponse'
        '400':
          $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - command
        - factory
      description: 'Delete a factory.'
      operationId: deleteFactory
      parameters:
        -
          name: id
          in: path
          description: 'Factory ID'
          required: true
          schema:
            type: integer
      responses:
        '200':
          $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/schemas/ErrorResponse'
        '412':
          $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/schemas/ErrorResponse'
  '/command/factory/{id}/storage':
    put:
      tags:
        - command
        - factory
      description: 'Update a factory storage.'
      operationId: updateFactoryStorage
      parameters:
        -
          name: id
          in: path
          description: 'Factory ID'
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              required:
                - materials
              properties:
                materials:
                  type: array
                  items: { required: [id, quantity], properties: { id: { description: 'Material ID.', type: integer }, quantity: { description: 'Quantity difference (positive for adding stock, negative for removing).', type: integer } }, type: object }
              type: object
      responses:
        '200':
          $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/schemas/ErrorResponse'
        '400':
          $ref: '#/components/schemas/ErrorResponse'
        '412':
          $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    FactoryCreateRequest:
      properties:
        name:
          type: string
        capacity:
          type: integer
          minimum: 1
      type: object
...
    FactoryUpdateRequest:
      properties:
        name:
          type: string
          nullable: true
        storageCapacity:
          type: integer
          minimum: 1
          nullable: true
      type: object
...
    ErrorResponse:
      properties:
        title:
          type: string
          example: 'Error title'
        type:
          $ref: '#/components/schemas/ErrorType'
        detail:
          type: string
          example: 'Error description'
          nullable: true
        exception:
          properties:
            message:
              type: string
              example: 'Some exception description'
            code:
              type: integer
              example: 123
            trace:
              type: array
              items:
                type: object
              example:
                -
                  file: index.php
                  line: 1
                  function: abc
                  args: ['Argument value']
          type: object
          nullable: true
        values:
          type: object
          example:
            key1: value1
            key2: value2
          nullable: true
      type: object
    SuccessResponse:
      properties:
        message:
          type: string
          example: Message
        detail:
          type: string
          example: Description
          nullable: true
        values:
          type: object
          example:
            key1: value1
            key2: value2
          nullable: true
      type: object
    ErrorType:
      type: string
      enum:
        - validation_error
        - database_error
        - internal_error
        - resource_not_found_error
        - resource_access_error
...